requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.24.4

torch==2.0.1
sentence-transformers==2.6.1
transformers==4.38.2

faiss-cpu==1.7.4

fastapi
uvicorn
pymongo
boto3
python-dotenv

spacy==3.7.2
pypdf
python-docx

tqdm
phonenumbers
pandas
requests
python-multipart
gunicorn
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
email-validator

# ✅ Google Drive (NEW)
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
